[provisioning] enable reentrant provisioning flow
The CP device ID is extacted from flash during CP and passed to FT. If
the CP step is skipped, e.g., if the device is already in test_locked*,
then host will not know what the CP device ID is.

This updates the provisioning flow to extract the CP device ID from
flash again during FT and pass it back to the host, making the flow
reentrant.

Signed-off-by: Tim Trippel <ttrippel@google.com>