[sival] Add option to skip reset on lc state read.
Add `opentitantool` argument to skip reset on `LcStateRead`. This is
useful when debugging device alerts  as the device will report
escalation state as a result of an escalation. Since the escalation
status is not persisted across resets, we need to be able to connect to
LC_TAP to read without performing a reset. This is only possible in lc
states that support continuous PINMUX_TAP polling.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>