[ottf,console] Move device-specific code out of print.c
Console/Device-specific code should not be in print.c which only contains
the printing related functionality. The console specific code is moved to
new file in the test_framework directory and are split between UART and SPI
for better code organization. For compatiblity, we cannot phase out
`base_uart_stdout` because it is used in quite a few SRAM program and
non-OTTF-based flash programs that may not want to include the full OTTF
console. Therefore this one is kept with a "minimal" UART driver. For now
the OTTF UART console code is duplicated but will soon evolve into its own
thing.

No code was changed in this PR, it was only moved/duplicated.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>