[entropy_src/rtl] Absorb full window before triggering SHA3 processing
It can happen that a health test window ends while the conditioner is
busy. Before this commit, the final processing of the conditioner got
triggered whenever testing the last symbol of a health test window and
independent of the fill status of the various FIFOs before the
conditioner. As a result, it could happen that the conditioner got
triggered before absorbing the last couple of symbols belonging to the
health test window.

To avoid this, this commit modifies the RTL to monitor the fill status
of the distribution and preconditioner FIFOs, and to delay the process
command accordingly.

This resolves lowRISC/OpenTitan#24085.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>