[opentitanlib, testutils] Handle multiple transfers in SPI bitbanging decoder
The current SPI bitbanging decoding logic assumed that there would only
be one transfer in the read samples, with CS low at the start. To support
adding improved error messages and decoding of variable-length SPI words
in the future, add support for arbitrary numbers of SPI transmissions.

Also take the opportunity to fix the size of `byte`: `u8` suffices.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>