[pentest] Bugfix csr_combi test
The CSR combi test writes values to working registers to read them out later.
The triggers in between the test can interfere with those. Instead, use the
volatile register variable in C in order to make it clear that those registers
should not be overwritten to make the test stable.

Signed-off-by: Siemen Dhooghe <sdhooghe@google.com>