[dv,dma] Interrupt coverage refinements
Separate coverage of interrupt configuration from the
coverage of interrupt clearing writes (addr/data).

Collect the latter from the TL-UL traffic that the DUT
produces to clear interrupts, as opposed to just the
generated configuration. This ensures that the DUT is
actually capable of driving out the addresses and data
values of interest.

Lose references to address and data channel within
dma_scoreboard.sv because strictly these are called A and D
in the TileLink specification.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>