[rstmgr] Avoid consistency check failures due to CDC instrumentation
The module is built such that the parent reset can never de-assert after
the child reset. However, depending on the simulation tool and its
settings, the CDC instrumentation may delay the synchronization of the
parent reset de-assertion until after the child reset de-assertion which
then triggers an unexpected consistency check failure. To avoid this,
we disable the CDC randomization for the parent reset assertion
synchronization.

This is related to lowRISC/OpenTitan#27659.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>