[rust] fix wrong rustfmt edition
We're using aspect_rules_lint to perform formatting, which is just
calling `rustfmt` directly -- this is invoking rustfmt with the wrong
edition -- it will think the source is edition 2015 instead of 2024.
This prevents syntax features exclusive to newer editions (e.g. async
functions, let chains).

I can't find a way to fix this in Bazel, so use .rustfmt.toml instead.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>