[opentitanlib] make `poll_read` the primitive for uart and console
Instead of using `read` and `read_timeout` plus `non_blocking_help`, this
will make `poll_read` becoming the primary way of interacting with UART.

It will only process whatever is in the buffer, and if there isn't any,
`cx` is to be notified once it does. For a few implementations that cannot
support notification, I added a helper function to replicate the polling behaviour.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>