[entropy_src/dv] Run block-level DV with actual FIFO depths
Previously, block-level DV was running with a distribution FIFO depth of
2 independent of the configuration. This size is not appropriate for
the Darjeeling configuration and leads to lots of entropy drops, causing
various block-level DV issues. This commit changes the simulation config
to use the same parameters used by the instantiations in Earlgrey and
Darjeeling.

Related to this, the commit also reduces the size of the esfinal FIFO
from 4 to 3. This results in an RTL but this doesn't have an impact on
the system. Both Earlgrey and Darjeeling are anyway instantiation the
ENTROPY_SRC with a esfinal FIFO depth of 3. Changing default parameter
just ensures we simulate the proper configuration in block-level DV.

This is related to lowRISC/OpenTitan#27471.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>