[entropy_src] Disable CS AES halt request interface for Darjeeling
This commit disables the CS AES halt request interface for Darjeeling as
the interface is not required. Disabling the interface reduces the
backpressure experienced by the hardware pipeline as the integrated
conditioner never has to wait for CSRNG's AES core to become idle.

To save area, this commit also reduces the depth of the distribution
FIFO which has the main purpose of absorbing incoming entropy in case
of backpressure.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>