[rv_timer,dv] Sort out spinwait timeout in rv_timer_random_vseq
There are two changes. Firstly, the spinwait had a bizarre extra wait
after reading the right value! Reorder things to be a bit more
sensible.

Secondly, I realised that if delay is large and spinwait_timeout_ns is
small, we need to add to the timeout so that we actually wait long
enough to read the right value.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>