[ci] Fix bug re: bitstream choice in FPGA tests
The "BITSTREAM" environment variable was not being passed through to
bitstreams_workspace.py, causing the test job to silently choose a bitstream
from the cloud cache over the one downloaded from GHA from the same workflow
run.

This commit fixes this defect, ensuring that the bitstream built by the CI in a
workflow is the one that workflow invocation's FPGA tests run on.

Signed-off-by: Ryan Torok <rtorok@zerorisc.com>